restful + json格式请求体

+++ models/dept.go
@@ -103,7 +103,7 @@ func (m *Dept) FindById(id int) (dept Dept, err error) {
 	return
 }
 
-func (m *Dept) FindByMap(offset, limit int, dataMap map[string]interface{},orderBy string) (res []Dept, total int, err error) {
+func (m *Dept) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []Dept, total int, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
